+++ exchangelib/queryset.py
@@ -468,6 +468,13 @@ class QuerySet(object):
         if self._cache is not None and not args and not kwargs:
             # We can only safely use the cache if get() is called without args
             items = self._cache
+        elif not args and set(kwargs.keys()) == {'item_id', 'changekey'}:
+            # We allow calling get(item_id=..., changekey=...) to get a single item, but only if exactly these two
+            # kwargs are present.
+            items = list(self.folder.fetch(
+                ids=[(kwargs['item_id'], kwargs['changekey'])],
+                only_fields=self.only_fields,
+            ))
         else:
             new_qs = self.filter(*args, **kwargs)
             items = list(new_qs.__iter__())